requirements.txt
# Core dependencies for the AI Test Case Generator project
streamlit>=1.38.0          # Web UI
groq>=0.9.0                # Groq API oficcial Client
python-dotenv>=1.0.1       # Safe .venv handling

# Future additions for ML-based test generation:
scikit-learn>=1.5.0        # ML (RandomForest, etc.)
pandas>=2.2.0              # Datasets handling
numpy>=1.26.0              # Common dependency of pandas/sklearn
